# backend/openapi.yaml
openapi: 3.0.3
info:
  title: KAYAD API
  description: |
    KAYAD - Kenya's Premium Car Marketplace API
    
    This API provides endpoints for managing car listings, auctions, bids, escrow transactions, and user accounts.
    
    ## Authentication
    
    Most endpoints require authentication using JWT tokens. Include the token in the Authorization header:
    
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    
    API requests are rate limited. Standard limits are 100 requests per minute per user.
    
    ## Versioning
    
    The API is versioned using URL path versioning (e.g., /api/v1/cars). Current version is v1.
    
    ## Deprecation Policy
    
    Deprecated endpoints will be supported for at least 6 months before removal.
  version: 1.0.0
  contact:
    name: KAYAD API Support
    email: api@kayad.space
    url: https://kayad.space
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.kayad.space/api/v1
    description: Production server
  - url: https://staging-api.kayad.space/api/v1
    description: Staging server
  - url: http://localhost:5000/api/v1
    description: Development server
...
paths:
...
  /cars:
    get:
      tags:
        - Cars
      summary: List cars
      description: Get paginated list of cars with optional filters
      operationId: listCars
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
        - name: status
          in: query
          schema:
            type: string
            enum: [active, sold, pending, inactive]
          description: Filter by status
        - name: minPrice
          in: query
          schema:
            type: number
          description: Minimum price
        - name: maxPrice
          in: query
          schema:
            type: number
          description: Maximum price
        - name: make
          in: query
          schema:
            type: string
          description: Filter by make
        - name: model
          in: query
          schema:
            type: string
          description: Filter by model
        - name: year
          in: query
          schema:
            type: integer
          description: Filter by year
        - name: location
          in: query
          schema:
            type: string
          description: Filter by location
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [createdAt, price, mileage, year]
            default: createdAt
          description: Sort field
        - name: sortOrder
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: List of cars
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarListResponse'
        '400':
          description: Invalid parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1m
        max: 100

    post:
      tags:
        - Cars
      summary: Create car listing
      description: Create a new car listing
      operationId: createCar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCarRequest'
      responses:
        '201':
          description: Car created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - seller account required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1h
        max: 50
  /cars/{carId}:
    get:
      tags:
        - Cars
      summary: Get car details
      description: Get detailed information about a specific car
      operationId: getCar
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: string
          description: Car ID
      responses:
        '200':
          description: Car details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1m
        max: 100

    put:
      tags:
        - Cars
      summary: Update car listing
      description: Update an existing car listing
      operationId: updateCar
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: string
          description: Car ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCarRequest'
      responses:
        '200':
          description: Car updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CarResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1h
        max: 50

    delete:
      tags:
        - Cars
      summary: Delete car listing
      description: Delete a car listing
      operationId: deleteCar
      parameters:
        - name: carId
          in: path
          required: true
          schema:
            type: string
          description: Car ID
      responses:
        '200':
          description: Car deleted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - not the owner
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Car not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1h
        max: 20
...
  /auctions/{auctionId}/bids:
    get:
      tags:
        - Bids
      summary: List auction bids
      description: Get all bids for an auction
      operationId: listAuctionBids
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: string
          description: Auction ID
        - name: page
          in: query
          schema:
            type: integer
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
          description: Items per page
      responses:
        '200':
          description: List of bids
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidListResponse'
        '404':
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1m
        max: 50

    post:
      tags:
        - Bids
      summary: Place bid
      description: Place a bid on an auction
      operationId: placeBid
      parameters:
        - name: auctionId
          in: path
          required: true
          schema:
            type: string
          description: Auction ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlaceBidRequest'
      responses:
        '201':
          description: Bid placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - auction not active
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Auction not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      security:
        - bearerAuth: []
      rateLimit:
        window: 1m
        max: 30
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
  schemas:
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Error message
        code:
          type: string
          description: Error code
        details:
          type: object
          description: Additional error details
    SuccessResponse:
      type: object
      required:
        - success
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    User:
      type: object
      required:
        - id
        - email
        - name
        - role
      properties:
        id:
          type: string
          description: User ID
        email:
          type: string
          format: email
          description: User email
        name:
          type: string
          description: User name
        phone:
          type: string
          description: User phone
        role:
          type: string
          enum: [user, seller, admin, inspector]
          description: User role
        emailVerified:
          type: boolean
          description: Email verification status
        profilePicture:
          type: string
          description: Profile picture URL
        location:
          type: string
          description: User location
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
    Car:
      type: object
      required:
        - id
        - title
        - price
        - mileage
        - year
        - make
        - model
        - status
      properties:
        id:
          type: string
          description: Car ID
        title:
          type: string
          description: Car title
        description:
          type: string
          description: Car description
        price:
          type: number
          description: Car price in KES
        mileage:
          type: number
          description: Car mileage in km
        year:
          type: integer
          description: Car year
        make:
          type: string
          description: Car make
        model:
          type: string
          description: Car model
        transmission:
          type: string
          enum: [automatic, manual]
          description: Transmission type
        fuelType:
          type: string
          description: Fuel type
        bodyType:
          type: string
          description: Body type
        color:
          type: string
          description: Car color
        location:
          type: string
          description: Car location
        images:
          type: array
          items:
            type: string
          description: Car image URLs
        features:
          type: array
          items:
            type: string
          description: Car features
        status:
          type: string
          enum: [active, sold, pending, inactive]
          description: Car status
        seller:
          $ref: '#/components/schemas/User'
        verification:
          type: object
          properties:
            verified:
              type: boolean
            verifiedBy:
              type: string
            verifiedAt:
              type: string
              format: date-time
          description: Verification status
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
        updatedAt:
          type: string
          format: date-time
          description: Update timestamp
    CreateCarRequest:
      type: object
      required:
        - title
        - price
        - mileage
        - year
        - make
        - model
        - images
      properties:
        title:
          type: string
          description: Car title
        description:
          type: string
          description: Car description
        price:
          type: number
          description: Car price in KES
        mileage:
          type: number
          description: Car mileage in km
        year:
          type: integer
          description: Car year
        make:
          type: string
          description: Car make
        model:
          type: string
          description: Car model
        transmission:
          type: string
          enum: [automatic, manual]
          description: Transmission type
        fuelType:
          type: string
          description: Fuel type
        bodyType:
          type: string
          description: Body type
        color:
          type: string
          description: Car color
        location:
          type: string
          description: Car location
        images:
          type: array
          items:
            type: string
          description: Car image URLs
        features:
          type: array
          items:
            type: string
          description: Car features
    UpdateCarRequest:
      type: object
      properties:
        title:
          type: string
          description: Car title
        description:
          type: string
          description: Car description
        price:
          type: number
          description: Car price in KES
        mileage:
          type: number
          description: Car mileage in km
        year:
          type: integer
          description: Car year
        make:
          type: string
          description: Car make
        model:
          type: string
          description: Car model
        transmission:
          type: string
          enum: [automatic, manual]
          description: Transmission type
        fuelType:
          type: string
          description: Fuel type
        bodyType:
          type: string
          description: Body type
        color:
          type: string
          description: Car color
        location:
          type: string
          description: Car location
        images:
          type: array
          items:
            type: string
          description: Car image URLs
        features:
          type: array
          items:
            type: string
          description: Car features
        status:
          type: string
          enum: [active, sold, pending, inactive]
          description: Car status
    CarResponse:
      type: object
      required:
        - car
      properties:
        car:
          $ref: '#/components/schemas/Car'
    CarListResponse:
      type: object
      required:
        - cars
        - pagination
      properties:
        cars:
          type: array
          items:
            $ref: '#/components/schemas/Car'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
...
    Bid:
      type: object
      required:
        - id
        - auctionId
        - userId
        - amount
        - timestamp
      properties:
        id:
          type: string
          description: Bid ID
        auctionId:
          type: string
          description: Auction ID
        userId:
          type: string
          description: User ID
        amount:
          type: number
          description: Bid amount in KES
        timestamp:
          type: string
          format: date-time
          description: Bid timestamp
    PlaceBidRequest:
      type: object
      required:
        - amount
      properties:
        amount:
          type: number
          description: Bid amount in KES
    BidResponse:
      type: object
      required:
        - bid
      properties:
        bid:
          $ref: '#/components/schemas/Bid'
    BidListResponse:
      type: object
      required:
        - bids
        - pagination
      properties:
        bids:
          type: array
          items:
            $ref: '#/components/schemas/Bid'
        pagination:
          type: object
          properties:
            page:
              type: integer
            limit:
              type: integer
            total:
              type: integer
            totalPages:
              type: integer
...